Higher quality hash function for AbstractInfo
Have a higher quality hash function for AbstractInfo.

For example, it is common for many AttrInfos (a subclass of AbstractInfo) to differ only by their path. The old hash function would hash these all to the same value, potentially causing considerable performance loss if they are all in the same HashMap.